requirements.txt
flask==2.0.1
flask-cors==3.0.10
pytesseract==0.3.10
pillow==9.0.1
pandas==1.3.5
sentence-transformers==2.2.2
scikit-learn==1.0.2
gunicorn==20.1.0